iOS: add platform services, MIDI scaffolding, and preset cleanup
- move background playback setup out of the app entry point and into shared audio/session services
- add now-playing and remote-command scaffolding plus CoreMIDI discovery/mapping foundations
- split native drum preset data into shared model files and retire the old monolithic preset source
- expand AppState, preset management, and main/preset views to surface native-only runtime and MIDI state
- refactor large SwiftUI control views for maintainability and update the iOS README to match the new platform scope

+++ KesshoiOS/Kessho/Audio/DrumSynth.swift
@@ -969,7 +969,9 @@ class DrumSynth {
                     if pattern[currentStep] {
                         // Probability check
                         if Double(rng()) <= lane.probability {
-                            let velocity = Float(lane.velocityMin + rng() * Float(lane.velocityMax - lane.velocityMin))
+                            let velocity = Float(
+                                lane.velocityMin + (Double(rng()) * (lane.velocityMax - lane.velocityMin))
+                            )
                             // Pick random voice from enabled voices
                             let selectedVoice = voices[Int(rng() * Float(voices.count)) % voices.count]
                             triggerVoice(selectedVoice, velocity: velocity * Float(lane.level))
